Add FP-smoothing evaluation, deferral-curve axis-matching, and PCA-weighted probe experiments
Multi-seed validation of FP-smoothing (K=5, beta=0.25) stacked on the
combined (min-pos + pct=0.3 catfilter) recipe on both EmbedLLM protocols:
helps modestly on unseen but hurts on all-seen, so excluded from the
official track. Also adds the pct=0.3 catfilter methodology confirmation,
ablation gap-fill tables, PCA-loading-weighted probe reallocation (V1.5),
deferral-curve plots axis-matched to the CSCR paper figure, and rejected
experiments (FP-refine, ensemble, prompt-level PCA) kept as documented
negative results.

+++ src/router/cost_models.py
@@ -8,8 +8,8 @@ def get_latency_ms(expert_id: str) -> float:
 def get_cost_per_1k(expert_id: str) -> float:
     return float(REGISTRY.get(expert_id, {}).get('cost_usd_per_1k_tokens', 0.0))
 
-def get_param_count(expert_id: str) -> int:
-    return int(REGISTRY.get(expert_id, {}).get('n_params', 0))
+def get_param_count(expert_id: str) -> float:
+    return float(REGISTRY.get(expert_id, {}).get('n_params', 0))
 
 def compute_cost(expert_id: str, n_tokens: int, cost_type="usd") -> float:
     if cost_type == "latency":
